+++ src/middlewares/index.ts
@@ -1,0 +1,2 @@
+export { SchemaValidationMiddleware } from './SchemaValidationMiddleware';
+export { BasicAuthMiddleware } from './BasicAuthMiddleware';