+++ src/utils/index.ts
@@ -15,4 +15,4 @@ export {
 
 export { loadConfigIni, loadDotEnv, loadYaml } from './config';
 
-export { mountURL } from './global';
+export { evaluateSchema, mountURL } from './global';